Exctract barChart fixtures from testData
